Move spec directories
